requirements.txt
requests
tweepy
python-whois